Move `<WordList/>` into own file
